fix(components): stats module extracted to stats-table
