+++ apps/artemis-web/components/bgp-table/bgp-table.spec.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 
-import BGPTable from './bgptable';
+import BGPTable from './bgp-table';
 
 describe('BGPTable', () => {
   it('should render successfully', () => {
