+++ apps/artemis-web/components/bgp-table/bgp-table.tsx
